update for project refs
