+++ apps/example/package.json
@@ -12,7 +12,7 @@
 		"start": "cross-env pnpm run build && electron ./dist/main.mjs",
 		"test": "echo 'only e2e'",
 		"test:e2e": "cross-env pnpm run build && playwright test",
-		"typecheck": "tsc",
+		"typecheck": "tsc --build --noEmit",
 		"lint": "oxlint --type-aware && eslint . && oxfmt . --check",
 		"typeinst": "tsc --extendedDiagnostics | grep Instantiations:"
 	},
